move helper functions into separate file
